cambios

+++ database/migrations/2021_03_28_153150_create_asignacion_table.php
@@ -25,7 +25,7 @@ class CreateAsignacionTable extends Migration
             $table->string('material');
             $table->float('precio_unitario');
             $table->string('unidad_medida');
-            $table->integer('capacidad');
+            $table->float('capacidad');
             $table->timestamps();
         });
     }
