+++ database/migrations/2021_03_28_153449_create_detalle_nota_asignacion_table.php
@@ -24,7 +24,7 @@ class CreateDetalleNotaAsignacionTable extends Migration
             $table->string('tipo_tanque');
             $table->string('material');
             $table->string('unidad_medida');
-            $table->integer('capacidad');
+            $table->float('capacidad');
             $table->timestamps();
         });
     }
